projects moved to client side
