--- browser/react/components/navbar/NavbarContainer.js
@@ -1,12 +1,0 @@
-import { connect } from 'react-redux';
-import Navbar from './Navbar';
-import { leaveProject } from '../../store/modules/selectedProject'
-
-const mapDispatchToProps = dispatch => ({
-  removeProject: () => dispatch(leaveProject())
-});
-
-export default connect(
-  null,
-  mapDispatchToProps
-)(Navbar);
