--- browser/react/components/projects/ProjectsContainer.js
@@ -1,11 +1,0 @@
-import { connect } from 'react-redux';
-import Projects from './Projects';
-
-const mapStateToProps = ({ projects, projectName }) => ({
-  projects,
-  fullView: !projectName
-})
-
-export default connect(
-  mapStateToProps
-)(Projects);
