+++ browser/react/store/modules/index.js
@@ -1,10 +1,6 @@
 import { combineReducers } from 'redux';
 import messageStatus from './contact';
-import projects from './projects';
-import projectName from './selectedProject';
 
 export default combineReducers({
-  messageStatus,
-  projects,
-  projectName
+  messageStatus
 });
