+++ browser/react/utils/constants.js
