--- browser/utils/index.js
@@ -1,4 +1,0 @@
-import { AMAZON_S3_BUCKET } from './constants'
-
-export const getStaticResourceLink = filename =>
-  `${AMAZON_S3_BUCKET}${filename}`
